VS Code: platform specific remote url getters
