+++ vscode/src/extension.common.ts
@@ -20,6 +20,7 @@ import type { ContextRankerConfig, ContextRankingController } from './local-cont
 import type { LocalEmbeddingsConfig, LocalEmbeddingsController } from './local-context/local-embeddings'
 import type { SymfRunner } from './local-context/symf'
 import { start } from './main'
+import type { RemoteUrlGetter } from './repository/repo-name-resolver'
 import type {
     OpenTelemetryService,
     OpenTelemetryServiceConfig,
@@ -43,6 +44,7 @@ export interface PlatformContext {
         | Constructor<typeof SourcegraphNodeCompletionsClient>
     createSentryService?: (config: Pick<ConfigurationWithAccessToken, 'serverEndpoint'>) => SentryService
     createOpenTelemetryService?: (config: OpenTelemetryServiceConfig) => OpenTelemetryService
+    getRemoteUrlGetters?: () => RemoteUrlGetter[]
     startTokenReceiver?: typeof startTokenReceiver
     onConfigurationChange?: (configuration: Configuration) => void
     extensionClient: ExtensionClient
